@opencv-alalek 2023-05-18
Please rebase to resolve conflicts:

```
Conflicting files
modules/dnn/src/dnn_common.hpp
modules/dnn/test/test_backends.cpp
```